Moved `file` class out from `plugins` directory
