+++ lib/file.js
@@ -3,7 +3,7 @@
 const fs              = require('fs-extra')
 const path            = require('upath')
 // TODO: Evaluate removal of BaseFile class
-const BaseFile        = require('../module/basefile')
+const BaseFile        = require('./module/basefile')
 
 module.exports = class File extends BaseFile {
