+++ lib/jus.js
@@ -23,9 +23,6 @@ module.exports = function jus (sourceDir, targetDir) {
   // This has to be synchronous
   // ... we do not want watcher to start without all file type handlers loaded
   fs.readdirSync(filesDir).forEach((file) => {
-    // Ignore file `file.js` (because it is not a type handler, it contains...
-    // ... parent class File) and files without '.js' extension
-    if (path.basename(file, '.js') === 'file' || path.extname(file) !== '.js') return
     sentinel.use(path.resolve(filesDir, file))
   })
