+++ lib/plugins/datafile.js
@@ -4,7 +4,7 @@ const fs              = require('fs')
 const path            = require('upath')
 const yaml            = require('js-yaml')
 const snakeCase       = require('lodash').snakeCase
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class datafile extends File {
   constructor(filepath, sourceDir, targetDir) {
