+++ lib/plugins/image.js
@@ -5,7 +5,7 @@ const path            = require('upath')
 const getImageColors  = require('get-image-colors')
 const exif            = require('exif-parser')
 const imageSize       = require('image-size')
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class Image extends File {
   constructor(filepath, sourceDir, targetDir) {
