+++ lib/plugins/layout.js
@@ -3,7 +3,7 @@
 const fs              = require('fs')
 const path            = require('upath')
 const frontmatter     = require('html-frontmatter')
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class Layout extends File {
