+++ lib/plugins/page.js
@@ -8,7 +8,7 @@ const frontmatter     = require('html-frontmatter')
 const handlebars      = require('handlebars')
 const lobars          = require('lobars')
 const titlecase       = require('inflection').titleize
-const File            = require('./file')
+const File            = require('../file')
 
 handlebars.registerHelper(lobars)
