+++ lib/plugins/script.js
@@ -7,7 +7,7 @@ const babelify        = require('babelify')
 const preset_es2015   = require('babel-preset-es2015')
 const preset_react    = require('babel-preset-react')
 const concat          = require('concat-stream')
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class Script extends File {
