+++ lib/plugins/stylesheetCSS.js
@@ -3,7 +3,7 @@
 const fs              = require('fs')
 const path            = require('upath')
 const myth            = require('myth')
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class StylesheetCSS extends File {
