+++ lib/plugins/stylesheetLESS.js
@@ -3,7 +3,7 @@
 const fs              = require('fs')
 const path            = require('upath')
 const less            = require('less')
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class StylesheetLESS extends File {
