+++ lib/plugins/stylesheetSASS.js
@@ -3,7 +3,7 @@
 const fs              = require('fs')
 const path            = require('upath')
 const sass            = require('node-sass')
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class StylesheetSASS extends File {
