+++ lib/plugins/stylesheetSTYL.js
@@ -3,7 +3,7 @@
 const fs              = require('fs')
 const path            = require('upath')
 const stylus          = require('stylus')
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class StylesheetSTYL extends File {
