+++ lib/plugins/unknown.js
@@ -2,7 +2,7 @@
 
 const fs              = require('fs')
 const path            = require('upath')
-const File            = require('./file')
+const File            = require('../file')
 
 module.exports = class Unknown extends File {
